refactor: moving UserPageContext.jsx from 'src/context/UserPageContext.jsx' to 'src/contexts/UserPageContext.jsx'
