+++ src/components/UserPage/CardDeleteButton.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Button from '@mui/material/Button';
 import { deleteItinerary } from '../../api/userpage';
-import { useItineraries } from '../../context/UserPageContext';
+import { useItineraries } from '../../contexts/UserPageContext';
 
 export default function CardDeleteButtons({id}) {
   const {itineraries, setItineraries} = useItineraries()
